move menu components into their own files
